service deleted imgs

+++ app/src/main/java/com/example/kittenappscollage/collect/DeletedImagesInList.java
@@ -25,13 +25,10 @@ import com.example.kittenappscollage.helpers.rx.ThreadTransformers;
 
 import java.io.FileNotFoundException;
 import java.util.Arrays;
-import java.util.Comparator;
 
 import io.reactivex.Observable;
 import io.reactivex.ObservableEmitter;
 import io.reactivex.ObservableOnSubscribe;
-import io.reactivex.functions.Action;
-import io.reactivex.functions.Consumer;
 
 import static com.example.kittenappscollage.helpers.Massages.LYTE;
 
@@ -60,9 +57,9 @@ public class DeletedImagesInList extends Service {
 
     private void startForegroundService() {
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            createNotificationChannel("my_service", "Delete files in device");
+            createNotificationChannel("my_service_delete", "Delete files in device");
         }
-        startForeground(2, buildNotification("my_service"));
+        startForeground(2, buildNotification("my_service_delete"));
     }
 
     @RequiresApi(api = Build.VERSION_CODES.O)
@@ -80,6 +77,8 @@ public class DeletedImagesInList extends Service {
         NotificationCompat.Builder builder =
                 new NotificationCompat.Builder(this, channelId)
                         .setSmallIcon(R.drawable.icon_delete_all)
+                        .setContentTitle("Удаление")
+                        .setContentText("Файлы удаляются с физического носителя")
                         .setShowWhen(true)
                         .setOngoing(true)
                         .setProgress(100, 0, true)
@@ -95,15 +94,15 @@ public class DeletedImagesInList extends Service {
             clearDeletedList(data,names,emitter);
             emitter.onComplete();
         }).compose(new ThreadTransformers.InputOutput<>())
-                .doOnComplete(() -> {
+          .doOnComplete(() -> {
                     stopForeground(true);
                     stopSelf();
                 })
-        .subscribe(pair -> {
+          .subscribe(pair -> {
                 if(!pair.second)
                 Massages.SHOW_MASSAGE(getApplicationContext(),
                         getApplicationContext().getResources().getString(R.string.ERROR_DELETE_FILE)+pair.first);
-
+//                else LYTE("delete - "+pair.first);
         });
     }
 
@@ -125,11 +124,17 @@ public class DeletedImagesInList extends Service {
 
     private int bruteForceOption(int index, DocumentFile[]files, String name,ObservableEmitter<Pair<String,Boolean>> emitter){
         for (int i=index;i<files.length;i++){
-            final String n = files[i].getName();
-            if(n.equals(name)){
-                index = i;
-                emitter.onNext(new Pair<>(name,deleteFile(files[i])));
-                break;
+            DocumentFile f = files[i];
+            if(f.isFile()) {
+                String type = f.getType();
+                if (type.equals("image/png") || type.equals("image/jpeg") || type.equals("image/jpg")) {
+                    final String n = f.getName();
+                    if (n != null && n.equals(name)) {
+                        index = i;
+                        emitter.onNext(new Pair<>(name, deleteFile(files[i])));
+                        break;
+                    }
+                }
             }
         }
         return index;
